Move Konstructs class to konstructs.cpp/h.
